Resolving problem with lifetime of UserDataViewModel.

+++ HardTrainingCore/Messages/ChangePageInCommonViewMessage.cs
@@ -2,13 +2,13 @@
 {
     public class ChangePageInCommonViewMessage
     {
-        public ChangePageInCommonViewMessage(TypesOfViews typeOfView, int profileId)
+        public ChangePageInCommonViewMessage(TypesOfViews typeOfView, short profileId)
         {
             this.TypeOfPageViewToOpen = typeOfView;
             this.ProfileId = profileId;
         }
 
         public TypesOfViews TypeOfPageViewToOpen { get; private set; }
-        public int ProfileId { get; private set; }
+        public short ProfileId { get; private set; }
     }
 }
